+++ src/components/Channel/Channel.tsx
@@ -168,6 +168,8 @@ export type ChannelPropsForwardedToComponentContext<
   | 'UnreadMessagesNotification'
   | 'UnreadMessagesSeparator'
   | 'VirtualMessage'
+  | 'StopAIGenerationButton'
+  | 'StreamedMessageText'
 >;
 
 const isUserResponseArray = <
@@ -1301,6 +1303,8 @@ const ChannelInner = <
       ReactionsList: props.ReactionsList,
       SendButton: props.SendButton,
       StartRecordingAudioButton: props.StartRecordingAudioButton,
+      StopAIGenerationButton: props.StopAIGenerationButton,
+      StreamedMessageText: props.StreamedMessageText,
       ThreadHead: props.ThreadHead,
       ThreadHeader: props.ThreadHeader,
       ThreadStart: props.ThreadStart,
@@ -1367,6 +1371,8 @@ const ChannelInner = <
       props.UnreadMessagesNotification,
       props.UnreadMessagesSeparator,
       props.VirtualMessage,
+      props.StopAIGenerationButton,
+      props.StreamedMessageText,
       props.emojiSearchIndex,
       props.reactionOptions,
     ],