+++ src/components/ChannelPreview/ChannelPreviewMessenger.tsx
@@ -1,5 +1,7 @@
 import React, { useRef } from 'react';
 import clsx from 'clsx';
+
+import { ChannelPreviewActionButtons as DefaultChannelPreviewActionButtons } from './ChannelPreviewActionButtons';
 import { Avatar as DefaultAvatar } from '../Avatar';
 import { useComponentContext } from '../../context';
 
@@ -7,9 +9,9 @@ import type { ChannelPreviewUIComponentProps } from './ChannelPreview';
 import type { DefaultStreamChatGenerics } from '../../types/types';
 
 const UnMemoizedChannelPreviewMessenger = <
-  StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics
+  SCG extends DefaultStreamChatGenerics = DefaultStreamChatGenerics
 >(
-  props: ChannelPreviewUIComponentProps<StreamChatGenerics>,
+  props: ChannelPreviewUIComponentProps<SCG>,
 ) => {
   const {
     active,
@@ -26,7 +28,11 @@ const UnMemoizedChannelPreviewMessenger = <
     watchers,
   } = props;
 
-  const { ChannelAvatar } = useComponentContext();
+  const {
+    ChannelAvatar,
+    ChannelPreviewActionButtons = DefaultChannelPreviewActionButtons,
+  } = useComponentContext<SCG>();
+
   const channelPreviewButton = useRef<HTMLButtonElement | null>(null);
 
   const avatarName =
@@ -46,44 +52,47 @@ const UnMemoizedChannelPreviewMessenger = <
   const Avatar = ChannelAvatar ?? PropsAvatar;
 
   return (
-    <button
-      aria-label={`Select Channel: ${displayTitle || ''}`}
-      aria-selected={active}
-      className={clsx(
-        `str-chat__channel-preview-messenger str-chat__channel-preview`,
-        active && 'str-chat__channel-preview-messenger--active',
-        unread && unread >= 1 && 'str-chat__channel-preview-messenger--unread',
-        customClassName,
-      )}
-      data-testid='channel-preview-button'
-      onClick={onSelectChannel}
-      ref={channelPreviewButton}
-      role='option'
-    >
-      <div className='str-chat__channel-preview-messenger--left'>
-        <Avatar
-          className='str-chat__avatar--channel-preview'
-          groupChannelDisplayInfo={groupChannelDisplayInfo}
-          image={displayImage}
-          name={avatarName}
-        />
-      </div>
-      <div className='str-chat__channel-preview-end'>
-        <div className='str-chat__channel-preview-end-first-row'>
-          <div className='str-chat__channel-preview-messenger--name'>
-            <span>{displayTitle}</span>
-          </div>
-          {!!unread && (
-            <div className='str-chat__channel-preview-unread-badge' data-testid='unread-badge'>
-              {unread}
+    <div className='str-chat__channel-preview-container'>
+      <ChannelPreviewActionButtons channel={channel} />
+      <button
+        aria-label={`Select Channel: ${displayTitle || ''}`}
+        aria-selected={active}
+        className={clsx(
+          `str-chat__channel-preview-messenger str-chat__channel-preview`,
+          active && 'str-chat__channel-preview-messenger--active',
+          unread && unread >= 1 && 'str-chat__channel-preview-messenger--unread',
+          customClassName,
+        )}
+        data-testid='channel-preview-button'
+        onClick={onSelectChannel}
+        ref={channelPreviewButton}
+        role='option'
+      >
+        <div className='str-chat__channel-preview-messenger--left'>
+          <Avatar
+            className='str-chat__avatar--channel-preview'
+            groupChannelDisplayInfo={groupChannelDisplayInfo}
+            image={displayImage}
+            name={avatarName}
+          />
+        </div>
+        <div className='str-chat__channel-preview-end'>
+          <div className='str-chat__channel-preview-end-first-row'>
+            <div className='str-chat__channel-preview-messenger--name'>
+              <span>{displayTitle}</span>
             </div>
-          )}
+            {!!unread && (
+              <div className='str-chat__channel-preview-unread-badge' data-testid='unread-badge'>
+                {unread}
+              </div>
+            )}
+          </div>
+          <div className='str-chat__channel-preview-messenger--last-message'>
+            {latestMessagePreview}
+          </div>
         </div>
-        <div className='str-chat__channel-preview-messenger--last-message'>
-          {latestMessagePreview}
-        </div>
-      </div>
-    </button>
+      </button>
+    </div>
   );
 };