+++ src/components/Message/index.ts
@@ -13,4 +13,5 @@ export * from './QuotedMessage';
 export * from './renderText';
 export * from './types';
 export * from './utils';
+export * from './StreamedMessageText';
 export type { TimestampProps } from './Timestamp';