+++ src/context/ComponentContext.tsx
@@ -6,6 +6,7 @@ import {
   AvatarProps,
   BaseImageProps,
   ChannelAvatarProps,
+  ChannelPreviewActionButtonsProps,
   CooldownTimerProps,
   CustomMessageActionsListProps,
   DateSeparatorProps,
@@ -38,6 +39,7 @@ import {
   RecordingPermissionDeniedNotificationProps,
   SendButtonProps,
   StartRecordingAudioButtonProps,
+  StreamedMessageTextProps,
   SuggestionItemProps,
   SuggestionListProps,
   ThreadHeaderProps,
@@ -61,6 +63,7 @@ import type {
   PropsWithChildrenOnly,
   UnknownType,
 } from '../types/types';
+import type { StopAIGenerationButtonProps } from '../components/MessageInput/StopAIGenerationButton';
 
 export type ComponentContextValue<
   StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics,
@@ -86,6 +89,10 @@ export type ComponentContextValue<
   BaseImage?: React.ComponentType<BaseImageProps>;
   /** Custom UI component to display channel avatar that may be different from a simple user avatar, defaults to and accepts same props as: [ChannelAvatar](https://github.com/GetStream/stream-chat-react/blob/master/src/components/Avatar/ChannelAvatar.tsx) */
   ChannelAvatar?: React.ComponentType<ChannelAvatarProps>;
+  /** Custom UI component to display set of action buttons within `ChannelPreviewMessenger` component, accepts same props as: [ChannelPreviewActionButtons](https://github.com/GetStream/stream-chat-react/blob/master/src/components/ChannelList/ChannelPreviewActionButtons.tsx) */
+  ChannelPreviewActionButtons?: React.ComponentType<
+    ChannelPreviewActionButtonsProps<StreamChatGenerics>
+  >;
   /** Custom UI component to display the slow mode cooldown timer, defaults to and accepts same props as: [CooldownTimer](https://github.com/GetStream/stream-chat-react/blob/master/src/components/MessageInput/CooldownTimer.tsx) */
   CooldownTimer?: React.ComponentType<CooldownTimerProps>;
   /** Custom UI component to render set of buttons to be displayed in the MessageActionsBox, defaults to and accepts same props as: [CustomMessageActionsList](https://github.com/GetStream/stream-chat-react/blob/master/src/components/MessageActions/CustomMessageActionsList.tsx) */
@@ -195,6 +202,8 @@ export type ComponentContextValue<
   SendButton?: React.ComponentType<SendButtonProps<StreamChatGenerics>>;
   /** Custom UI component button for initiating audio recording, defaults to and accepts same props as: [StartRecordingAudioButton](https://github.com/GetStream/stream-chat-react/blob/master/src/components/MediaRecorder/AudioRecorder/AudioRecordingButtons.tsx) */
   StartRecordingAudioButton?: React.ComponentType<StartRecordingAudioButtonProps>;
+  StopAIGenerationButton?: React.ComponentType<StopAIGenerationButtonProps> | null;
+  StreamedMessageText?: React.ComponentType<StreamedMessageTextProps>;
   /** Custom UI component that displays thread's parent or other message at the top of the `MessageList`, defaults to and accepts same props as [MessageSimple](https://github.com/GetStream/stream-chat-react/blob/master/src/components/Message/MessageSimple.tsx) */
   ThreadHead?: React.ComponentType<MessageProps<StreamChatGenerics>>;
   /** Custom UI component to display the header of a `Thread`, defaults to and accepts same props as: [DefaultThreadHeader](https://github.com/GetStream/stream-chat-react/blob/master/src/components/Thread/Thread.tsx) */